+++ nodalia-cards-2.0.4.js
@@ -1,2 +1,2 @@
-import "./nodalia-cards-2.0.5.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.0.5","contentSha256_12":"2b23652f4905","file":"nodalia-cards-2.0.4.js","targetFile":"nodalia-cards-2.0.5.js","fallbackFile":"nodalia-cards.js"};}
+import "./nodalia-cards-2.1.0.js";
+if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.0","contentSha256_12":"9a8644c61f40","file":"nodalia-cards-2.0.4.js","targetFile":"nodalia-cards-2.1.0.js","fallbackFile":"nodalia-cards.js"};}