+++ nodalia-cards.manifest.js
@@ -1,23 +1,23 @@
 export default {
-  "pkgVersion": "2.0.5",
-  "contentSha256_12": "2b23652f4905",
+  "pkgVersion": "2.1.0",
+  "contentSha256_12": "9a8644c61f40",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "compatLoaderFiles": [
-    "nodalia-cards-2.0.3.js",
-    "nodalia-cards-2.0.4.js"
+    "nodalia-cards-2.0.4.js",
+    "nodalia-cards-2.0.5.js"
   ],
-  "splitCoreFile": "nodalia-cards-core-2.0.5.js",
-  "splitCoreSha256_12": "5af8856172e6",
-  "splitSuiteFile": "nodalia-cards-suite-2.0.5.js",
-  "splitSuiteSha256_12": "cd80157e9e8a",
-  "editorFile": "nodalia-cards-editor-2.0.5.js",
-  "editorSha256_12": "cf952c17c474"
+  "splitCoreFile": "nodalia-cards-core-2.1.0.js",
+  "splitCoreSha256_12": "a5716f80a534",
+  "splitSuiteFile": "nodalia-cards-suite-2.1.0.js",
+  "splitSuiteSha256_12": "a1c2b9d2c242",
+  "editorFile": "nodalia-cards-editor-2.1.0.js",
+  "editorSha256_12": "7f50f338e82a"
 };
-export const pkgVersion = "2.0.5";
-export const contentSha256_12 = "2b23652f4905";
+export const pkgVersion = "2.1.0";
+export const contentSha256_12 = "9a8644c61f40";
 export const file = "nodalia-cards.bundle.js";
-export const splitCoreFile = "nodalia-cards-core-2.0.5.js";
-export const splitSuiteFile = "nodalia-cards-suite-2.0.5.js";
-export const editorFile = "nodalia-cards-editor-2.0.5.js";
+export const splitCoreFile = "nodalia-cards-core-2.1.0.js";
+export const splitSuiteFile = "nodalia-cards-suite-2.1.0.js";
+export const editorFile = "nodalia-cards-editor-2.1.0.js";